code for the daemon to support feedback of current mosaic panel and the option to image selected subset of panels. Two class parameters are exposed for class Seestar: cur_mosaic_nRA and cur_mosaic_nDec
The start_mosaic command will now take an optional parameter: "selected_panels". Example:

"selected_panels":"12;22"

Note that the session_time_sec will still have the original meaning as if all the panels are to be acquired. The reason is i expect this patient is called if some of the panels failed to be captured, so only the missing ones will be recaptured, with all the other parameters stay the same.

I also discovered a potential source of why autofocus fails during mosaic. If the is_use_autofocus parameter was missing from the request, the logic hangs. That has been fixed.