Change container to be based on ubuntu 24.04 (to fix build)
It seems that the current python code is trying to import `NotRequired` from the `typing` module, which is not available in Python 3.10.
The `NotRequired` type was introduced in Python 3.11.  Therefore use the 24.04 ubuntu as the basis for the container.  24.04 is the next LTS
after 22.04 and it has Python 3.12.

Traceback (most recent call last):
  File "/home/seestar/seestar_alp/root_app.py", line 13, in <module>
    from front.app import FrontMain, get_live_status
  File "/home/seestar/seestar_alp/front/app.py", line 45, in <module>
    from device import telescope
  File "/home/seestar/seestar_alp/device/telescope.py", line 26, in <module>
    from device.seestar_device import Seestar
  File "/home/seestar/seestar_alp/device/seestar_device.py", line 11, in <module>
    from typing import Optional, Any, TypedDict, NotRequired
ImportError: cannot import name 'NotRequired' from 'typing' (/usr/lib/python3.10/typing.py)